added product type and generalised product flow
