+++ client/types/Product.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    id: number;
+    category: string;
+    name: string;
+    description: string;
+    price: string;
+    image: string;
+}
